# 8332603: [PPC64] Improve itable_stub