--- a/src/hotspot/cpu/ppc/macroAssembler_ppc.cpp
+++ b/src/hotspot/cpu/ppc/macroAssembler_ppc.cpp
@@ -1864,6 +1864,132 @@ void MacroAssembler::cmpxchgd(ConditionRegister flag,
   // (flag == eq) => (dest_current_value == compare_value), ( swapped)
 }
 
+// Look up the method for a megamorphic invokeinterface call in a single pass over itable:
+// - check recv_klass (actual object class) is a subtype of resolved_klass from CompiledICHolder
+// - find a holder_klass (class that implements the method) vtable offset and get the method from vtable by index
+// The target method is determined by <holder_klass, itable_index>.
+// The receiver klass is in recv_klass.
+// On success, the result will be in method_result, and execution falls through.
+// On failure, execution transfers to the given label.
+void MacroAssembler::lookup_interface_method_stub(Register recv_klass,
+                                                  Register holder_klass,
+                                                  Register resolved_klass,
+                                                  Register method_result,
+                                                  Register temp_itbl_klass,
+                                                  Register scan_temp,
+                                                  int itable_index,
+                                                  Label& L_no_such_interface) {
+  // 'method_result' is only used as output register at the very end of this method.
+  // Until then, we can reuse it as 'holder_offset'.
+  Register holder_offset = method_result;
+
+  assert_different_registers(resolved_klass, recv_klass, holder_klass, temp_itbl_klass, scan_temp, holder_offset);
+
+  BLOCK_COMMENT("lookup_interface_method_stub {");
+
+  int vtable_start_offset = in_bytes(Klass::vtable_start_offset());
+  int itable_offset_entry_size = itableOffsetEntry::size() * wordSize;
+  int ioffset = in_bytes(itableOffsetEntry::interface_offset());
+  int ooffset = in_bytes(itableOffsetEntry::offset_offset());
+
+  Label L_loop_search_resolved_entry, L_resolved_found, L_holder_found;
+
+  lwz(scan_temp, in_bytes(Klass::vtable_length_offset()), recv_klass);
+  addi(recv_klass, recv_klass, vtable_start_offset + ioffset);
+
+  // itableOffsetEntry[] itable = recv_klass + Klass::vtable_start_offset() + sizeof(vtableEntry) * recv_klass->_vtable_len;
+  // temp_itbl_klass = itable[0]._interface;
+  int vtblEntrySize = vtableEntry::size_in_bytes();
+  sldi(scan_temp, scan_temp, exact_log2(vtblEntrySize));
+
+  ld(temp_itbl_klass, scan_temp, recv_klass);
+
+  li(holder_offset, 0);
+
+  // scan_temp = &(itable[0]._interface)
+  add(scan_temp, scan_temp, recv_klass);
+
+  // Initial checks:
+  //   - if (holder_klass != resolved_klass), go to "scan for resolved"
+  //   - if (itable[0] == holder_klass), shortcut to "holder found"
+  //   - if (itable[0] == 0), no such interface
+  cmpd(CCR0, holder_klass, resolved_klass);
+  bne(CCR0, L_loop_search_resolved_entry);
+
+  cmpd(CCR0, temp_itbl_klass, holder_klass);
+  beq(CCR0, L_holder_found);
+
+  cmpdi(CCR0, temp_itbl_klass, 0);
+  beq(CCR0, L_no_such_interface);
+
+  // Loop: Look for holder_klass record in itable
+  //   do {
+  //     scan_temp += itable_offset_entry_size
+  //     temp_itbl_klass = *(scan_temp);
+  //     if (temp_itbl_klass == holder_klass) {
+  //       goto holder_found; // Found!
+  //     }
+  //   } while (temp_itbl_klass != 0);
+  //   goto no_such_interface // Not found.
+  Label L_search_holder;
+  bind(L_search_holder);
+  addi(scan_temp, scan_temp, itable_offset_entry_size);
+  ld(temp_itbl_klass, 0, scan_temp);
+  cmpd(CCR0, holder_klass, temp_itbl_klass);
+  beq(CCR0, L_holder_found);
+  cmpdi(CCR0, temp_itbl_klass, 0);
+  bne(CCR0, L_search_holder);
+
+  b(L_no_such_interface);
+
+  // Loop: Look for resolved_class record in itable
+  //   while (true) {
+  //     scan_temp += itable_offset_entry_size
+  //     temp_itbl_klass = *(scan_temp);
+  //     if (temp_itbl_klass == 0) {
+  //       goto L_no_such_interface;
+  //     }
+  //     if (temp_itbl_klass == resolved_klass) {
+  //        goto L_resolved_found;  // Found!
+  //     }
+  //     if (temp_itbl_klass == holder_klass) {
+  //        holder_offset = scan_temp;
+  //     }
+  //   }
+  //
+  Label L_loop_search_resolved;
+  bind(L_loop_search_resolved);
+  addi(scan_temp, scan_temp, itable_offset_entry_size);
+  ld(temp_itbl_klass, 0, scan_temp);
+
+  bind(L_loop_search_resolved_entry);
+
+  cmpdi(CCR0, temp_itbl_klass, 0);
+  beq(CCR0, L_no_such_interface);
+
+  cmpd(CCR0, temp_itbl_klass, resolved_klass);
+  bne(CCR0, L_resolved_found);
+
+  cmpd(CCR0, temp_itbl_klass, holder_klass);
+  bne(CCR0, L_loop_search_resolved);
+
+  mr_if_needed(holder_offset, scan_temp);
+  b(L_loop_search_resolved);
+
+  // See if we already have a holder klass. If not, go and scan for it.
+  bind(L_resolved_found);
+  cmpdi(CCR0, holder_offset, 0);
+  beq(CCR0, L_search_holder);
+  mr_if_needed(scan_temp, holder_offset);
+
+  // Finally, scan_temp contains holder_klass vtable offset
+  bind(L_holder_found);
+  lwz(method_result, ooffset - ioffset, scan_temp);
+  addi(recv_klass, recv_klass, itable_index * wordSize + in_bytes(itableMethodEntry::method_offset())
+        - vtable_start_offset - ioffset); // subtract offsets to restore the original value of recv_klass
+  ld(method_result, method_result, recv_klass);
+}
+
 // Look up the method for a megamorphic invokeinterface call.
 // The target method is determined by <intf_klass, itable_index>.
 // The receiver klass is in recv_klass.