--- a/src/hotspot/cpu/ppc/macroAssembler_ppc.cpp
+++ b/src/hotspot/cpu/ppc/macroAssembler_ppc.cpp
@@ -1895,7 +1895,7 @@ void MacroAssembler::lookup_interface_method_stub(Register recv_klass,
   Label L_loop_search_resolved_entry, L_resolved_found, L_holder_found;
 
   lwz(scan_temp, in_bytes(Klass::vtable_length_offset()), recv_klass);
-  addi(recv_klass, recv_klass, vtable_start_offset + ioffset);
+  add_const_optimized(recv_klass, recv_klass, vtable_start_offset + ioffset);
 
   // itableOffsetEntry[] itable = recv_klass + Klass::vtable_start_offset() + sizeof(vtableEntry) * recv_klass->_vtable_len;
   // temp_itbl_klass = itable[0]._interface;
@@ -1933,7 +1933,7 @@ void MacroAssembler::lookup_interface_method_stub(Register recv_klass,
   //   goto no_such_interface // Not found.
   Label L_search_holder;
   bind(L_search_holder);
-  addi(scan_temp, scan_temp, itable_offset_entry_size);
+  add_const_optimized(scan_temp, scan_temp, itable_offset_entry_size);
   ld(temp_itbl_klass, 0, scan_temp);
   cmpd(CCR0, holder_klass, temp_itbl_klass);
   beq(CCR0, L_holder_found);
@@ -1959,7 +1959,7 @@ void MacroAssembler::lookup_interface_method_stub(Register recv_klass,
   //
   Label L_loop_search_resolved;
   bind(L_loop_search_resolved);
-  addi(scan_temp, scan_temp, itable_offset_entry_size);
+  add_const_optimized(scan_temp, scan_temp, itable_offset_entry_size);
   ld(temp_itbl_klass, 0, scan_temp);
 
   bind(L_loop_search_resolved_entry);
@@ -1973,19 +1973,19 @@ void MacroAssembler::lookup_interface_method_stub(Register recv_klass,
   cmpd(CCR0, temp_itbl_klass, holder_klass);
   bne(CCR0, L_loop_search_resolved);
 
-  mr_if_needed(holder_offset, scan_temp);
+  mr(holder_offset, scan_temp);
   b(L_loop_search_resolved);
 
   // See if we already have a holder klass. If not, go and scan for it.
   bind(L_resolved_found);
   cmpdi(CCR0, holder_offset, 0);
   beq(CCR0, L_search_holder);
-  mr_if_needed(scan_temp, holder_offset);
+  mr(scan_temp, holder_offset);
 
   // Finally, scan_temp contains holder_klass vtable offset
   bind(L_holder_found);
   lwz(method_result, ooffset - ioffset, scan_temp);
-  addi(recv_klass, recv_klass, itable_index * wordSize + in_bytes(itableMethodEntry::method_offset())
+  add_const_optimized(recv_klass, recv_klass, itable_index * wordSize + in_bytes(itableMethodEntry::method_offset())
         - vtable_start_offset - ioffset); // subtract offsets to restore the original value of recv_klass
   ld(method_result, method_result, recv_klass);
 }