@offamitkumar 2024-06-16
Testing (Power 8 - ppc-le): 

```
==============================                                                                                                          
Test summary                                                                                                                            
==============================                                                                                                          
   TEST                                              TOTAL  PASS  FAIL ERROR                                                            
>> jtreg:test/hotspot/jtreg:tier1                     2470  2469     1     0 <<                                                         
>> jtreg:test/jdk:tier1                               2413  2412     1     0 <<                                                         
>> jtreg:test/langtools:tier1                         4534  4533     1     0 <<                                                         
   jtreg:test/jaxp:tier1                                 0     0     0     0                                                            
   jtreg:test/lib-test:tier1                            33    33     0     0                                                            
==============================                                                                                                          
TEST FAILURE
```

```
root@crampon1:~/amit/jdk# cat $(find . -name newfailures.txt)                                                                           
# newfailures.txt                                                                                                                       
# newfailures.txt                                                                                                                       
gtest/GTestWrapper.java                                                                                                                 
# newfailures.txt                                                                                                                       
java/util/ResourceBundle/Control/MissingResourceCauseTestRun.java                                                                       
# newfailures.txt                                                                                                                       
jdk/javadoc/doclet/testIOException/TestIOException.java                                                                                 
# newfailures.txt                                                                                                                       
# newfailures.txt 
```
 two test are failing with `root` user; There are already issues on JBS: [TestIOException.java](https://bugs.openjdk.org/browse/JDK-8334195) & [MissingResourceCauseTestRun.java](https://bugs.openjdk.org/browse/JDK-8334333)

`GTestWrapper.java` is failing due to unspecified `-nativepath`;