@varada1110 2024-06-17
```
There are no related failures in tier1 with release and fastdebug on aix-ppc. Here is the benchmark results.

Before patch: 

Benchmark                        Mode  Cnt   Score   Error  Units
InterfaceCalls.test1stInt2Types  avgt   12  18.084 ? 0.160  ns/op
InterfaceCalls.test1stInt3Types  avgt   12  26.789 ? 0.282  ns/op
InterfaceCalls.test1stInt5Types  avgt   12  26.926 ? 0.438  ns/op
InterfaceCalls.test2ndInt2Types  avgt   12  18.500 ? 0.215  ns/op
InterfaceCalls.test2ndInt3Types  avgt   12  30.996 ? 0.489  ns/op
InterfaceCalls.test2ndInt5Types  avgt   12  28.395 ? 0.215  ns/op
InterfaceCalls.testIfaceCall     avgt   12  26.749 ? 0.238  ns/op
InterfaceCalls.testIfaceExtCall  avgt   12  28.189 ? 0.395  ns/op
InterfaceCalls.testMonomorphic   avgt   12  15.251 ? 0.126  ns/op
Finished running test 'micro:vm.compiler.InterfaceCalls'
Test report is stored in build/aix-ppc64-server-release/test-results/micro_vm_compiler_InterfaceCalls

==============================
Test summary
==============================
   TEST                                              TOTAL  PASS  FAIL ERROR   
   micro:vm.compiler.InterfaceCalls                      1     1     0     0   
==============================
TEST SUCCESS

Finished building target 'test' in configuration 'aix-ppc64-server-release'


After patch :

Benchmark                        Mode  Cnt   Score   Error  Units
InterfaceCalls.test1stInt2Types  avgt   12  18.078 ? 0.133  ns/op
InterfaceCalls.test1stInt3Types  avgt   12  26.816 ? 0.276  ns/op
InterfaceCalls.test1stInt5Types  avgt   12  26.918 ? 0.233  ns/op
InterfaceCalls.test2ndInt2Types  avgt   12  18.674 ? 0.204  ns/op
InterfaceCalls.test2ndInt3Types  avgt   12  30.360 ? 0.251  ns/op
InterfaceCalls.test2ndInt5Types  avgt   12  28.356 ? 0.394  ns/op
InterfaceCalls.testIfaceCall     avgt   12  26.676 ? 0.188  ns/op
InterfaceCalls.testIfaceExtCall  avgt   12  27.613 ? 0.258  ns/op
InterfaceCalls.testMonomorphic   avgt   12  15.175 ? 0.090  ns/op
Finished running test 'micro:vm.compiler.InterfaceCalls'
Test report is stored in build/aix-ppc64-server-release/test-results/micro_vm_compiler_InterfaceCalls

==============================
Test summary
==============================
   TEST                                              TOTAL  PASS  FAIL ERROR   
   micro:vm.compiler.InterfaceCalls                      1     1     0     0   
==============================
TEST SUCCESS

```