@offamitkumar 2024-06-17
>Testing is fine with release VM, I see one failure test/hotspot/jtreg/gtest/GTestWrapper.java with fastdebug on aix-ppc.

@varada1110 can you check the reason for the failure ? is it just because `-nativepath` is not specified ?